requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.2
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cytoolz==0.12.3
dataclasses-json==0.6.7
Deprecated==1.2.14
dune_client==1.7.7
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.8
iniconfig==2.0.0
isort==6.0.0
markdown-it-py==3.0.0
marshmallow==3.23.0
mccabe==0.7.0
mdformat==0.7.22
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
packaging==24.1
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==15.1.0
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
six==1.16.0
tomli==2.0.2
toolz==0.12.1
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20241003
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20240712
types-setuptools==75.2.0.20241025
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
web3==7.2.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.9.8
